* re-organized file structure * added apache configuration file * added install script * renamed everything from emily to mialb
