+++ MiaLB/Mia.wsgi
@@ -3,5 +3,4 @@ import sys
 
 print sys.path
 
-from api_router import api_router as application
 #from . import *
