+++ __init__.py
@@ -1,5 +1,3 @@
 #! /usr/bin/python
 
-from api_router import api_router
-
 __all__ = ['api_router']
