+++ conf/appache-mia-lb.conf
@@ -1,0 +1,21 @@
+WSGISocketPrefix /var/run/wsgi
+Listen 666
+
+<VirtualHost *:666>
+	WSGIDaemonProcess Mia-LB threads=16 python-path=/software/Mia/LB
+	WSGIScriptAlias / /software/Mia/LB/Mia.wsgi
+
+	<Directory /software/Mia/LB>
+		WSGIProcessGroup Mia-LB
+		WSGIApplicationGroup %{GLOBAL}
+		Options Indexes MultiViews FollowSymLinks
+		<IfVersion < 2.3 >
+			Order allow,deny
+			Allow from all
+		</IfVersion>
+		<IfVersion >= 2.3>
+			Require all granted
+		</IfVersion>
+	</Directory>
+
+</VirtualHost>
