+++ conf/emily.te
@@ -15,17 +15,9 @@ require {
 
 #!!!! This avc is allowed in the current policy
 allow httpd_t default_t:dir read;
-
-#!!!! This avc is allowed in the current policy
 allow httpd_t default_t:file { read getattr };
-
-#!!!! This avc is allowed in the current policy
 allow httpd_t hi_reserved_port_t:tcp_socket name_bind;
 allow httpd_t httpd_sys_script_exec_t:dir remove_name;
-
-#!!!! This avc is allowed in the current policy
 allow httpd_t httpd_sys_script_exec_t:dir { write add_name };
 allow httpd_t httpd_sys_script_exec_t:file unlink;
-
-#!!!! This avc is allowed in the current policy
 allow httpd_t httpd_sys_script_exec_t:file { write create };
